+++ CHANGELOG.md
@@ -322,6 +322,8 @@ v0.30.0 (2022-12-20)
 
 ### Bugfixes
 
+- Fix issue where an empty `server:` config stanza would cause debug-level logging. (@neomantra)
+
 - Fix issue where whitespace was being sent as part of password when using a
   password file for `redis_exporter`. (@spartan0x117)